Add namespace to accomodate /api/v1 in routes
