+++ backend/config/routes.rb
@@ -1,5 +1,9 @@
 Rails.application.routes.draw do
   get "up" => "rails/health#show", as: :rails_health_check
 
-  resources :subscribers
+  namespace :api do
+    namespace :v1 do
+      resources :subscribers
+    end
+  end
 end
